requirements.txt
openai-whisper
librosa
soundfile